# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ClassifiedPage": {
        "description": "A page of the board plus the count the UI needs to say \"Show 124 places\".\n\n`total` is the size of the whole result set, not of this page: a filter\npanel that cannot show how many places match forces people to paginate to\nfind out whether a filter did anything.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ClassifiedOut"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "ClassifiedPage",
        "type": "object"
      },
      "ContactRevealOut": {
        "properties": {
          "contact_phone": {
            "title": "Contact Phone",
            "type": "string"
          },
          "offer_id": {
            "title": "Offer Id",
            "type": "string"
          }
        },
        "required": [
          "offer_id",
          "contact_phone"
        ],
        "title": "ContactRevealOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Managed hospitality platform. This spec is generated from the implementation (code-first) and kept in sync by a CI drift-guard — see docs/api/README.md.",
    "title": "Homies API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/classifieds": {
      "get": {
        "description": "Search the free board. Only active offers, and never a phone number.",
        "operationId": "list_classifieds_v1_classifieds_get",
        "parameters": [
          {
            "in": "query",
            "name": "city",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "City"
            }
          },
          {
            "in": "query",
            "name": "district",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "District"
            }
          },
          {
            "in": "query",
            "name": "max_monthly_total",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Monthly Total"
            }
          },
          {
            "in": "query",
            "name": "min_monthly_total",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Monthly Total"
            }
          },
          {
            "in": "query",
            "name": "min_rooms",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Rooms"
            }
          },
          {
            "in": "query",
            "name": "min_area_m2",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Area M2"
            }
          },
          {
            "in": "query",
            "name": "furnished",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Furnished"
            }
          },
          {
            "in": "query",
            "name": "parking",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parking"
            }
          },
          {
            "in": "query",
            "name": "pets_allowed",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pets Allowed"
            }
          },
          {
            "in": "query",
            "name": "has_elevator",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Has Elevator"
            }
          },
          {
            "in": "query",
            "name": "available_by",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Available By"
            }
          },
          {
            "in": "query",
            "name": "max_term_months",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Term Months"
            }
          },
          {
            "in": "query",
            "name": "has",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Has"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "newest",
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassifiedPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Classifieds",
        "tags": [
          "properties"
        ]
      }
    },
    "/v1/classifieds/{offer_id}": {
      "get": {
        "operationId": "get_classified_v1_classifieds__offer_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "offer_id",
            "required": true,
            "schema": {
              "title": "Offer Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassifiedOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Classified",
        "tags": [
          "properties"
        ]
      }
    },
    "/v1/classifieds/{offer_id}/contact": {
      "post": {
        "description": "Disclose the owner's phone number to a verified account, and record it.\n\nA *verified phone* is the gate, not merely sign-in. An address costs a\nscraper nothing; a SIM costs money, and `users.phone` is unique, so the\ncost is paid per account rather than once. Every number taken still leaves\na row naming who took it. Anonymous and unverified visitors get the message\nchannel instead.\n\nEmail verification is deliberately NOT a second gate here: it adds a round\ntrip for the tenant and nothing an automated collector cannot buy in bulk.\nTightening that is a product call, not a technical gap.",
        "operationId": "reveal_contact_v1_classifieds__offer_id__contact_post",
        "parameters": [
          {
            "in": "path",
            "name": "offer_id",
            "required": true,
            "schema": {
              "title": "Offer Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactRevealOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reveal Contact",
        "tags": [
          "properties"
        ]
      }
    },
    "/v1/classifieds/{offer_id}/pause": {
      "post": {
        "operationId": "pause_classified_v1_classifieds__offer_id__pause_post",
        "parameters": [
          {
            "in": "path",
            "name": "offer_id",
            "required": true,
            "schema": {
              "title": "Offer Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassifiedOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Pause Classified",
        "tags": [
          "properties"
        ]
      }
    },
    "/v1/classifieds/{offer_id}/publish": {
      "post": {
        "operationId": "publish_classified_v1_classifieds__offer_id__publish_post",
        "parameters": [
          {
            "in": "path",
            "name": "offer_id",
            "required": true,
            "schema": {
              "title": "Offer Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassifiedOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Publish Classified",
        "tags": [
          "properties"
        ]
      }
    },
...
    "/v1/me/reveal-quota": {
      "get": {
        "description": "What the app shows before someone runs into the wall.\n\nA limit nobody can see is indistinguishable from a broken button.",
        "operationId": "my_reveal_quota_v1_me_reveal_quota_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevealQuotaOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "My Reveal Quota",
        "tags": [
          "properties"
        ]
      }
    },
...
  },
...
}